+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/system/service/impl/SysTenantServiceImpl.java
@@ -16,6 +16,10 @@ import org.springframework.stereotype.Service;
 import java.util.Collection;
 import java.util.List;
 
+/**
+ * @Description: 租户实现类
+ * @author: jeecg-boot
+ */
 @Service("sysTenantServiceImpl")
 @Slf4j
 public class SysTenantServiceImpl extends ServiceImpl<SysTenantMapper, SysTenant> implements ISysTenantService {
@@ -33,7 +37,7 @@ public class SysTenantServiceImpl extends ServiceImpl<SysTenantMapper, SysTenant
     }
 
     @Override
-    public int countUserLinkTenant(String id) {
+    public Long countUserLinkTenant(String id) {
         LambdaQueryWrapper<SysUser> userQueryWrapper = new LambdaQueryWrapper<>();
         userQueryWrapper.eq(SysUser::getRelTenantIds, id);
         userQueryWrapper.or().like(SysUser::getRelTenantIds, "%," + id);
@@ -46,7 +50,7 @@ public class SysTenantServiceImpl extends ServiceImpl<SysTenantMapper, SysTenant
     @Override
     public boolean removeTenantById(String id) {
         // 查找出已被关联的用户数量
-        int userCount = this.countUserLinkTenant(id);
+        Long userCount = this.countUserLinkTenant(id);
         if (userCount > 0) {
             throw new JeecgBootException("该租户已被引用，无法删除！");
         }
